News_Pipeline/News_API_Builder_2/News_API_Builder_2/News_Page_App_2/views.py
from rest_framework import generics
from django_filters.rest_framework import DjangoFilterBackend
from .models import News
from .serializer import NewsSerializer
from .filters import NewsFilter

class NewsList(generics.ListCreateAPIView):
    queryset = News.objects.all()
    serializer_class = NewsSerializer
    filter_backends = [DjangoFilterBackend]
    filterset_class = NewsFilter
